mypage component 분리
