--- src/styles/mypage.css
@@ -1,87 +1,0 @@
-.my-container {
-  border: #000000 1px solid;
-  /* sidebar의 너비를 고려한 왼쪽 여백 */
-  margin-right: 4%;
-  margin-bottom: 5%;
-  margin-left: 220px;
-  padding: 20px;
-  flex-grow: 1;
-}
-
-/* 
-.my-map-header {
-  width: 105.89px;
-  height: 33.04px;
-  left: 400.9px;
-  top: 222px;
-
-  font-family: "Inter";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 20.331px;
-  line-height: 25px;
-  display: flex;
-  align-items: center;
-
-  color: #000000;
-
-  margin: 0 0 0 5px;
-} */
-
-#map {
-  width: 70%;
-  height: 70vh;
-
-  margin-top: 2%;
-  margin-left: 15%;
-
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.my-map-section:nth-last-child(2) {
-  position: relative;
-  margin-left: 0;
-}
-
-.my-map-upload-button {
-  background-color: #000000;
-  border: none;
-  color: white;
-  padding: 15px 32px;
-  text-align: center;
-  text-decoration: none;
-  display: inline-block;
-  font-size: 16px;
-  cursor: pointer;
-  border-radius: 4px;
-  border: #000000 1px solid;
-}
-
-.my-map-upload-button:hover {
-  background-color: #ffffff;
-  color: black;
-  border: #000000 1px solid;
-}
-
-.my-folder-container {
-  border: #000000 1px solid;
-  /* sidebar의 너비를 고려한 왼쪽 여백 */
-  margin-bottom: 5%;
-  padding: 20px;
-}
-
-.my-folder-wrap {
-  width: 1000px;
-
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-}
-
-.my-map-readytravel-wrap {
-  border: #000000 1px solid;
-}
